Factor out server option processing into a separate module
The server has grown a fair bit of options and related validation logic,
so split the option parsing code into a separate module.  This also
makes it easier to reference the `ServerSettings` tuple from other
modules.
